pyproject.toml
[project]
name = "azure-functions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "openai>=1.88.0",
    "pymupdf>=1.26.1",
    "python-docx>=1.2.0",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "azure-functions",
]

[tool.uv.sources]
azure-functions = { workspace = true }
